# doc/电商平台.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "电商平台",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/product/showList": {
      "get": {
        "summary": "获取商品列表",
        "x-apifox-folder": "商品管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "查询所有商品并展示，对没有数据有异常抛出判断",
        "tags": [
          "商品管理"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "price": {
                            "type": "number"
                          },
                          "isList": {
                            "type": "boolean"
                          },
                          "shop": {
                            "type": "string"
                          },
                          "category": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "description",
                          "price",
                          "isList",
                          "category",
                          "shop"
                        ],
                        "x-apifox-orders": [
                          "id",
                          "name",
                          "description",
                          "price",
                          "isList",
                          "shop",
                          "category"
                        ],
                        "x-apifox-ignore-properties": []
                      },
                      "nullable": true
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "devMessages": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data",
                    "errors",
                    "devMessages"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "message",
                    "data",
                    "errors",
                    "devMessages"
                  ],
                  "x-apifox-ignore-properties": []
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "success": true,
                      "message": null,
                      "data": [
                        {
                          "id": 1,
                          "name": "api",
                          "description": "修改",
                          "price": 618,
                          "isList": false,
                          "shop": {
                            "id": 1,
                            "shopName": "test1"
                          },
                          "category": {
                            "id": 1,
                            "categoryName": "test1"
                          }
                        }
                      ],
                      "errors": [],
                      "devMessages": []
                    }
                  }
                }
              }
            }
          },
          "x-200:暂无商品": {
            "description": "暂无商品",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    },
                    "data": {
                      "type": "string",
                      "nullable": true
                    },
                    "errors": {
                      "type": "string"
                    },
                    "devMessages": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data",
                    "errors",
                    "devMessages"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "message",
                    "data",
                    "errors",
                    "devMessages"
                  ],
                  "x-apifox-ignore-properties": []
                },
                "examples": {
                  "1": {
                    "summary": "异常示例",
                    "value": {
                      "success": false,
                      "message": "没有数据",
                      "data": null,
                      "errors": [],
                      "devMessages": []
                    }
                  }
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/2902459/apis/api-90619004-run"
      }
    },
    "/product/findByName": {
      "get": {
        "summary": "根据商品名搜索",
        "x-apifox-folder": "商品管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "根据输入的商品名进行模糊查询，对未找到商品或相似商品有异常抛出判断\n将使用SQL语句的`like`\n",
        "tags": [
          "商品管理"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "",
            "required": true,
            "example": "api",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "模糊查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "price": {
                            "type": "integer"
                          },
                          "isList": {
                            "type": "boolean"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "description",
                          "price",
                          "isList"
                        ],
                        "x-apifox-orders": [
                          "id",
                          "name",
                          "description",
                          "price",
                          "isList"
                        ],
                        "x-apifox-ignore-properties": []
                      },
                      "nullable": true
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "devMessages": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data",
                    "errors",
                    "devMessages"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "message",
                    "data",
                    "errors",
                    "devMessages"
                  ],
                  "x-apifox-ignore-properties": []
                },
                "examples": {
                  "1": {
                    "summary": "模糊查询成功示例",
                    "value": {
                      "success": true,
                      "message": null,
                      "data": [
                        {
                          "id": 2,
                          "name": "test3",
                          "description": "好",
                          "price": 100,
                          "isList": false
                        },
                        {
                          "id": 5,
                          "name": "test1",
                          "description": "test",
                          "price": 0,
                          "isList": false
                        },
                        {
                          "id": 6,
                          "name": "test2",
                          "description": "test",
                          "price": 0,
                          "isList": false
                        }
                      ],
                      "errors": [],
                      "devMessages": []
                    }
                  }
                }
              }
            }
          },
          "x-200:单个查询成功": {
            "description": "单个查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "price": {
                          "type": "integer"
                        },
                        "isList": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "description",
                        "price",
                        "isList"
                      ],
                      "x-apifox-orders": [
                        "id",
                        "name",
                        "description",
                        "price",
                        "isList"
                      ],
                      "x-apifox-ignore-properties": [],
                      "nullable": true
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "nullable": true
                      }
                    },
                    "devMessages": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data",
                    "errors",
                    "devMessages"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "message",
                    "data",
                    "errors",
                    "devMessages"
                  ],
                  "x-apifox-ignore-properties": []
                },
                "examples": {
                  "1": {
                    "summary": "单个查询成功示例",
                    "value": {
                      "success": true,
                      "message": null,
                      "data": {
                        "id": 1,
                        "name": "test",
                        "description": "test",
                        "price": 0,
                        "isList": false
                      },
                      "errors": [],
                      "devMessages": []
                    }
                  }
                }
              }
            }
          },
          "x-200:查询失败": {
            "description": "查询失败",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "devMessages": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data",
                    "errors",
                    "devMessages"
                  ],
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": [
                    "success",
                    "message",
                    "data",
                    "errors",
                    "devMessages"
                  ]
                },
                "examples": {
                  "1": {
                    "summary": "异常示例",
                    "value": {
                      "success": false,
                      "message": "未找到该商品或相似商品",
                      "data": null,
                      "errors": [],
                      "devMessages": []
                    }
                  }
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/2902459/apis/api-90619014-run"
      }
    },
    "/product/new": {
      "post": {
        "summary": "增加商品",
        "x-apifox-folder": "商品管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "按以下请求参数要求输入商品属性，添加商品，对商品已存在、商家不存在、分类不存在有异常抛出判断",
        "tags": [
          "商品管理"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "",
            "required": true,
            "example": "api",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "price",
            "in": "query",
            "description": "",
            "required": true,
            "example": "850",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isList",
            "in": "query",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "description",
            "in": "query",
            "description": "",
            "required": true,
            "example": "test",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shop_id",
            "in": "query",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    },
                    "data": {
                      "type": "null"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "devMessages": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data",
                    "errors",
                    "devMessages"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "message",
                    "data",
                    "errors",
                    "devMessages"
                  ],
                  "x-apifox-ignore-properties": []
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "success": true,
                      "message": null,
                      "data": null,
                      "errors": [],
                      "devMessages": []
                    }
                  }
                }
              }
            }
          },
          "x-200:失败": {
            "description": "失败",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    },
                    "data": {
                      "type": "null"
                    },
                    "errors": {
                      "type": "string"
                    },
                    "devMessages": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data",
                    "errors",
                    "devMessages"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "message",
                    "data",
                    "errors",
                    "devMessages"
                  ],
                  "x-apifox-ignore-properties": []
                },
                "examples": {
                  "1": {
                    "summary": "异常示例",
                    "value": {
                      "success": false,
                      "message": "商品'api'已存在",
                      "data": null,
                      "errors": [],
                      "devMessages": []
                    }
                  }
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/2902459/apis/api-90619393-run"
      }
    },
...
  },
...
  "servers": [
    {
      "url": "localhost:8080",
      "description": "测试环境"
    }
  ]
}